Update pong.ino
Added the setup(). It begins by uses a TV.begin function, where NTSC indicates use for north american televisions.
Then the position of the pong ball is set to the middle of the screen.  
After this the user gets to choose both the speed of the pong ball, and then the play to score for the game.